updated @DeleteMapping

+++ Backend/MODULE2/2.5/putpatchanddelete/src/main/java/com/example/putpatchanddelete/services/EmployeeService.java
@@ -59,8 +59,12 @@ public class EmployeeService {
 
     }
 
-    public void deleteEmployeeById(Long employeeId) {
+    public boolean deleteEmployeeById(Long employeeId) {
+
+        boolean exist = employeeRepository.existsById(employeeId);
+        if(!exist) return false;
         employeeRepository.deleteById(employeeId);
+        return true;
 
 
     }
